Convert `byrne-thesis.docx` to `byrne-thesis.html` using bootstrap.
Previously, "byrne-thesis" was only available in proprietrary formats (PDF, DOC, DOCX), was not responsive, and had significant accessibility issues, not to mention, was not in a web-first format.

By running through an automated conversion process and adding Twitter Bootstrap header and footer, "byrne-thesis" can now be served natively as HTML, is fully responsive, and fully 508 compatabible.